chore(structure): co-locate home components, rename components/pages to components/shell, move projects header, update imports
