+++ app/(home)/_components/hero/config.ts
