+++ app/(home)/_hooks/useScrollCssVariables.ts
