+++ app/projects/_components/AnimatedHeader.client.tsx
